+++ docs/changelog.md
@@ -2,6 +2,9 @@
 
 ## Current
 
+* feat: Write empty permission list as <none> for Group Perm updates (ASG-3387)
+* fix: Fix missing variable in MSAAD log
+
 ## v1.5.0
 
 * feat: Update authaus to v1.0.36